Buidling a good response view

+++ app/entities/responses/UserResponse.scala
@@ -2,4 +2,4 @@ package app.entities.responses
 
 import app.utils._
 
-case class UserResponse(id:BigInt,status:StatusEnums.Value,firstname:String,lastname:String,role:RoleResponse,email:String,timezone:String,locale:String,avatar_url:String,phone_number:String)
+case class UserResponse(id:Int,status:StatusEnums.Value,firstname:String,lastname:String,role:RoleResponse,email:String,timezone:String,locale:String,avatar_url:String,phone_number:String)
